Wait for Delancey to finish saving
Doing this in a goroutine causes the client to trigger the delete, so
then Delancey ends up doing both save and delete at the same time which
causes a bunch of weird issues, one main one is save or delete trying to
use the container data after it's been removed from Delancey causing nil
dereference panics. Another is data committed can go missing because
it's removed before commit can finish.

This will probably be reverted once we complete the progress stuff since
that can be used to see when the save is complete, but the save always
has to finish before the delete can start.